Fix inconsistent Intel hardware intrinsic APIs

+++ src/Common/src/CoreLib/System/Runtime/Intrinsics/X86/Avx2.PlatformNotSupported.cs
@@ -1787,7 +1787,7 @@ namespace System.Runtime.Intrinsics.X86
         /// __m256i _mm256_sad_epu8 (__m256i a, __m256i b)
         ///   VPSADBW ymm, ymm, ymm/m256
         /// </summary>
-        public static Vector256<ulong> SumAbsoluteDifferences(Vector256<byte> left, Vector256<byte> right) { throw new PlatformNotSupportedException(); }
+        public static Vector256<ushort> SumAbsoluteDifferences(Vector256<byte> left, Vector256<byte> right) { throw new PlatformNotSupportedException(); }
 
         /// <summary>
         /// __m256i _mm256_unpackhi_epi8 (__m256i a, __m256i b)
